+++ notification-service/build.gradle
@@ -34,6 +34,8 @@ dependencies {
 
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
 
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+
     //common
     implementation project(':common')
